Add a color measurement.
Detail: To know the proposed color for each accepted move (mainly insertion
and removal), I have utilized the alps measurement tool to keep track of it.
The result is extracted in the evaluate_basics function.

To do: Because during the thermalization sweeps no measurement is performed,
strictly speaking this way of measuring color is flawed. In a future update,
I will define a global vector to keep track of this information (just like
how "nprop" and "nacc" are recorded) so that one won't miss anything.