Efficiency improved; minor revision.
1. The update efficiency (accepted rate/proposed rate) is slightly
improved by removing the random selection of colors BEFORE update.
For inserting, the color is determined on the fly (with equal probabilities);
for removal, one just selects the (anti)segment to be removed and
checks if its end colors are identical.

2. The output message when initializing Delta is modified.

Potential problem: when N_ENV=2, the insertion and removal acception
rate is slightly unbalanced (about 0.05%). Not sure if this matters...