global color count added
In the previous commit, a color measurement based on the ALPS utility
was introduced; however, it does not the counts during thermalization
sweeps, so the number does not match the acceptance rates printed out
in the end of simulation. By adding a global vector "ncolor" (similar
to "nacc" and "nprop") this problem is solved.

For now both approaches are used, but in the future version the ALPS
color measurement should be depreciated.