1. algorithm update (still buggy!); 2. segment output modified
Because there are colors in the segments when n_env > 1, the number of segments
and antisegments at order k is not necessary k. Therefore, functions to count
these numbers are added in local_configuration and used in the update. But this
is still buggy...

The operator<< of the segment class is modified.