1. algorithm update (bug seems fixed!) 2. new check consistency function
1. The functions now work as expected (at the programming level; the physics
is to be verified). No sign problem.

2. To help debug when modifying the algorithm, a function called check_n_segments_consistency
is added to the local_configuration class. It will use an alternative way to
compute the number of colored segments and antisegments, and compare with the
fast approach used in the algorithm tweak. Because this function seems to be
time consuming, it only works when VERY_VERBOSE is turned on.