Complete! Passed the compiler check and can be executed.
The ohmic environment is now added to the code as part of the local_configuration
class, because in this way it is easier to access the start and end times of segments.
The weight change contributed by the ohmic environment is computed as combination
of phase correlators (see the note for detail).

By default, the ohmic environment is turned off. To turn it on, set Dissipation=1.
Two additional parameters are required: r (the dissipation strength) and C0 (the
capacitance ratio of the 0-th color to the total capacitance). The code will give
an error message if r or C0 is not specified.

If N_ENV=2, C1 will be automatically computed as 1-C0.

The formula of the phase correlator, J(tau), is given in the note, in which the
cutoff frequency is set as 100/beta.

TODO:
1. when N_ENV=1, C0 can only be 1 (check with Harold!), so a quick check should
   be added.
2. J(tau) may not have the constant term (Euler constant).
3. The cutoff frequency may need to be changed or be given by the user.