The dissipation part is now a standalone class.
The disspation codes that were part of the local_configuration class are
now moved into a new, standalone class called "dissipation_configuration".
Unlike the local and hyb classes, this class stores *nothing* about the
dissipative environment, since everything (mainly the weight change) can
be computed on the fly. The purpose of this class is just to make the code
more structured and readable. In particular, one can think of this class
as not existing at all when Dissipation=0 is set.