1. measure_G modified; 2. small changes.
1. In the presence of the dissipative environment, the local Green's function
and its higher-order relatives will need to be modified. Fortunately, this
contribution is precisely the part that the dissipation contributes to the
weight change, which is already computed. Therefore, it is easy to include this
factor.

The modification is done for the imaginary-time Green's function now. Other
local correlations need to be modified later too.

2. Now the dissipation_configuration class is not a friend of local_configuration,
only its dissipation_weight_change function is.