DEBUGGING not finshed: rescaling the Green function
The local Green's function is discontinuous at end points, which are given by local
density measurements. However, from the raw data stored in the HDF5 file, in which
the end points are not replaced yet, I saw that if G_M(tau) is rescaled such that

          G(tau) = G_M(tau) / 2|G_M(0)+G_M(beta)|,

where the subscript M means the measured data, and replacing the end points after
the rescaling, then the resulting Green's function seems to become continuous. This
version is testing this idea.

Note that this rescaling is NOT YET justified on a physical ground!