conductance measurement is add; extensive tests REQUIRED!!!
The linear conductance for a single-impurity Anderson model can be expressed
as \sigma(iwn) = \int_0^\beta d\tau 2 cos(wn \tau) G(tau) \Delta(\beta-\tau)
in Matsubara frequency (see my note). Because the impurity Green's function
G(\tau) can be measured using CTHYB, the conductance can also be measured (in
the same way as how G(iwn) is measured).

In the current implementation, the conductance measurement is a member function
of the hybconfig class, as it's easier to access the hybridization data \Delta.
Because the contribution to conductance for a given color (i.e., lead) includes
all orbitals (i.e., spin), the measurement is performed separately for each
color and orbital, and then the result is post-processed in the evaluation stage
where measurements are assumed independent for the moment. The final outcome,
labeled by giwn_i, is the conductance for color i.

Note that the iwn->0 limit should be taken to obtain the true conductance; see
Mahan. This step requires careful extrapolation and thus is not implemented in
the code. In addition, the number of Matsubara frequency is again given by N_W.

Usage:
In the input file, set "MEASURE_conductance=1" (turned off by default) and
specify "N_W". (This measurement is independent of MEASURE_freq or MEASURE_time!)

TODO:
1. extend the formula of \sigma(iwn) to include the contribution of dissipation.
2. examine the raw data to make sure the post-processing procedure is correct,
   including check the measurements for different orbitals are independent.
3. examine the result and compare with the literature to check the validaty of
   this measurement.
4. try to measure G(\tau) only and do numerical integration to compare the outcome
   of two different (but should-be-equivalent) approaches.