"sign issue" for two leads fixed
For one lead (N_ENV=1), if the worm head is wrapped around (say go from
the left to the right of the tail), then no additional sign is needed,
otherwise both weight_ratio_ and permutation_sign_ need a -1.

For two leads (N_ENV=2), in addition to the rule discussed above, one
also needs to count the "imbalance" (difference between the number of c
and c^dagger of a particular color, modulo 2) within the worm body
(excluding the head and tail) that also contributes to the sign.