color_swap_update added
By setting COLORSWAP=1 in the input, the color_swap_update is performed.
It is a local update which picks a pair of c and cdagger (with the same
color) in a randomly chosen orbital and changes their color.

This commit seems to be bug-free, but further tests are needed!!!!!

TODO: think about a better way to avoid the ugly hack for generating
random numbers...