bug fix and minor changes; see the change log below
1. permutation factor computed for color_swap_update

The proposal probability was neglected in the previous implementation,
which is added now. Observation: without taking the proposal probability
into account, the updated colors would be imbalanced even if the coupling
strength for both colors is the same; that is, one color would get
updated more frequently than the other.

During the color swap update, one needs to randomly pick a c and a
cdagger, so the proposal probability for swapping out of a given color
with matrix size n is 1/n^2. The proposal probability for the inverse
move is then 1/(nbar+1)^2 (nbar being the matrix size for the other color
BEFORE the update), so the permutation factor is n^2/(nbar+1)^2.

TODO: do more checks to confirm this idea!!!

2. disable all *n_segment* related functions as they are useless
3. disable worm updates, color flip update, and local and global spin flip
updates
4. code clean up and organized