bug fix and minor changes
1. fix permutation_sign_ issue in color_swap_update

When swapping the color of a remote pair of c and cdagger, the number of
"left-alone" operators with the same color between the pair needs to be
taken into account in order to get the sign correct. The guideline is to
keep permutation_sign_ = time_ordering_sign_ when the matrix is
rebuilded in time order.

TODO: do more checks to confirm this

2. allow color_flip_update
3. changes the way of counting color updates in color_flip_update