Dockerfile modified and wrapper provided
The wrapper is for local build only, as the ADD/COPY mechanism in the
Dockerfile will inevitably expose the not-yet-open-sourced source. The
workaround is to squash the layers during docker build (so that rm takes
effect), and then push the built image to Docker Hub. The Automated
Build there should not be deployed until the source is made public.